+++ modulus/utils/mesh/__init__.py
@@ -16,3 +16,4 @@
 
 from .combine_vtp_files import combine_vtp_files
 from .convert_file_formats import convert_tesselated_files_in_directory
+from .generate_stl import sdf_to_stl
